+++ CHANGELOG.md
@@ -148,7 +148,7 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 
 * `super/i-block/modules/field`:
   * Fixed working with watchers based on accessors
-  * Fixed resolving of accessor
+  * Fixed resolving of accessors
 
 ## v3.0.0-rc.90 (2020-10-22)